// tests/Resources/Data/Signup_Data.ts
export type Data = {
  title: string
  password: string
  day: string
  month: string
  year: string
  firstName: string
  lastName: string
  country: string
  address: string
  state: string
  city: string
  zipcode: string
  mobile: string
}

export const signupData: Data = {
  title: 'Mr.',
  password: '123456',
  day: '16',
  month: '11',
  year: '1998',
  firstName: 'Le',
  lastName: 'Kien',
  country: 'United States',
  address: 'NewYork',
  state: 'abc',
  city: 'HN',
  zipcode: '123',
  mobile: '123'
}